requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_s3_assets
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_ssm
aws_cdk.aws_route53
aws_cdk.aws_cloudwatch_actions
aws_cdk.aws_cloudwatch
aws_cdk.aws_lambda
aws_cdk.aws_apigateway
aws_cdk.aws_budgets